Extract utils related to RatingMetalColor out to utils/RatingMetalColor.ts. Resolve reviews.
