+++ atcoder-problems-frontend/src/components/DifficultyCircle.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import { Badge, Tooltip } from "reactstrap";
-import { CompletedRatingColor, getRatingColor } from "../utils";
+import { getRatingColor } from "../utils";
+import { RatingColorWithMetal } from "../utils/RatingMetalColor";
 import ProblemModel, {
   isProblemModelWithDifficultyModel,
   isProblemModelWithTimeModel,
@@ -26,7 +27,7 @@ interface LocalState {
   tooltipOpen: boolean;
 }
 
-function getColor(difficulty: number): CompletedRatingColor {
+function getColor(difficulty: number): RatingColorWithMetal {
   if (difficulty >= 3200) {
     if (difficulty < 3600) {
       return "Bronze";
